// src/enums/TextType.ts
enum TextType {
    BLOCKED_ACCOUNT = 'BLOCKED_ACCOUNT',
    WITHOUT_FUNDS = 'WITHOUT_FUNDS',
    ONLY_ONE_PURCHASE = 'ONLY_ONE_PURCHASE',
    WRONG_ACCESS_DATA = 'WRONG_ACCESS_DATA',
    WRONG_VERIFICATION_CODE = 'WRONG_VERIFICATION_CODE',
    ADD_FUNDS_TITLE = 'ADD_FUNDS_TITLE',
    ADD_FUNDS_DESCRIPTION = 'ADD_FUNDS_DESCRIPTION',
    WHITDRAW_FUNDS_TITLE = 'WHITDRAW_FUNDS_TITLE',
    WHITDRAW_FUNDS_DESCRIPTION = 'WHITDRAW_FUNDS_DESCRIPTION',
    TRANSACTION_SUCCESS = 'TRANSACTION_SUCCESS',
    TRANSACTION_ERROR = 'TRANSACTION_ERROR',
    AUCTION_START = 'AUCTION_START',
    AUCTION_CLOSE = 'AUCTION_CLOSE',
    TICKET_GENERATED = 'TICKET_GENERATED',
    INVOICE_GENERATED = 'INVOICE_GENERATED',
    USER_NOT_FOUND = 'USER_NOT_FOUND'
}
export default TextType;